refactor: refactoring discord implementation
